Move plot button to ConnectionResultsScreen
